+++ src/API/Reports/DataStore.php
@@ -16,7 +16,7 @@ use \Automattic\WooCommerce\Admin\API\Reports\TimeInterval;
 /**
  * Admin\API\Reports\DataStore: Common parent for custom report data stores.
  */
-class DataStore {
+class DataStore extends SqlQuery {
 
 	/**
 	 * Cache group for the reports.
@@ -72,20 +72,6 @@ class DataStore {
 	 * @var string
 	 */
 	private $context = 'reports';
-	/**
-	 * List of SQL clauses.
-	 *
-	 * @var array
-	 */
-	private $sql_clauses = array(
-		'select'     => array(),
-		'from'       => array(),
-		'outer_from' => array(),
-		'where'      => array(),
-		'where_time' => array(),
-		'limit'      => array(),
-		'order_by'   => array(),
-	);
 
 	/**
 	 * Class constructor
@@ -115,57 +101,6 @@ class DataStore {
 	 */
 	protected function prepend_table_name( $query, $field_name ) {
 		return str_replace( $field_name, $this->get_db_table_name() . '.' . $field_name, $query );
-	}
-
-	/**
-	 * Add a SQL clause to be included when get_data is called.
-	 *
-	 * @param string $clause SQL clause.
-	 * @param string $type   (select|from|outer_from|where|where_time|order_by|limit).
-	 */
-	protected function add_sql_clause( $clause, $type ) {
-		if ( isset( $this->sql_clauses[ $type ] ) && ! empty( $clause ) ) {
-			$this->sql_clauses[ $type ][] = $clause;
-		}
-	}
-
-	/**
-	 * Get SQL clause by type.
-	 *
-	 * @param string $type Clause type (select|from|outer_from|where|where_time|order_by|limit).
-	 *
-	 * @return string SQL clause.
-	 */
-	protected function get_sql_clause( $type ) {
-		if ( ! isset( $this->sql_clauses[ $type ] ) ) {
-			return '';
-		}
-
-		/**
-		 * Filter SQL clauses by type and context.
-		 *
-		 * @param array  $clauses The original arguments for the request.
-		 * @param string $context The data store context.
-		 */
-		$clauses = apply_filters( "wc_admin_clauses_{$type}", $this->sql_clauses[ $type ], $this->context );
-		/**
-		 * Filter SQL clauses by type and context.
-		 *
-		 * @param array  $clauses The original arguments for the request.
-		 */
-		$clauses = apply_filters( "wc_admin_clauses_{$type}_{$this->context}", $clauses );
-		return implode( ' ', $clauses );
-	}
-
-	/**
-	 * Clear SQL clauses by type.
-	 *
-	 * @param string $type Clause type (select|from|outer_from|where|where_time|order_by|limit).
-	 */
-	protected function clear_sql_clause( $type ) {
-		if ( isset( $this->sql_clauses[ $type ] ) ) {
-			$this->sql_clauses[ $type ] = array();
-		}
 	}
 
 	/**
@@ -494,11 +429,10 @@ class DataStore {
 			$intervals_query['where_time_clause'] .= " AND {$table_name}.date_created >= '$adj_after'";
 			$intervals_query['limit']              = 'LIMIT 0,' . $intervals_query['per_page'];
 			// @todo remove passed parameter assignment after down stream classes are updated.
-			$this->clear_sql_clause( 'where_time' );
-			$this->add_sql_clause( " AND {$table_name}.date_created <= '$adj_before'", 'where_time' );
-			$this->add_sql_clause( " AND {$table_name}.date_created >= '$adj_after'", 'where_time' );
-			$this->clear_sql_clause( 'limit' );
-			$this->add_sql_clause( 'LIMIT 0,' . $intervals_query['per_page'], 'limit' );
+			$this->clear_sql_clause( array( 'where_time', 'limit' ) );
+			$this->add_sql_clause( 'where_time', " AND {$table_name}.date_created <= '$adj_before'" );
+			$this->add_sql_clause( 'where_time', " AND {$table_name}.date_created >= '$adj_after'" );
+			$this->add_sql_clause( 'limit', 'LIMIT 0,' . $intervals_query['per_page'] );
 		} else {
 			if ( 'asc' === $query_args['order'] ) {
 				$offset = ( ( $query_args['page'] - 1 ) * $intervals_query['per_page'] ) - ( $expected_interval_count - $db_interval_count );
@@ -512,7 +446,7 @@ class DataStore {
 				$intervals_query['limit'] = 'LIMIT ' . $offset . ',' . $count;
 				// @todo remove passed parameter assignment after down stream classes are updated.
 				$this->clear_sql_clause( 'limit' );
-				$this->add_sql_clause( 'LIMIT ' . $offset . ',' . $count, 'limit' );
+				$this->add_sql_clause( 'limit', 'LIMIT ' . $offset . ',' . $count );
 			}
 			// Otherwise no change in limit clause.
 			$query_args['adj_after']  = $query_args['after'];
@@ -685,9 +619,7 @@ class DataStore {
 			'where_clause'      => '',
 		);
 		// @todo remove $sql_query after down stream classes are updated.
-		$this->clear_sql_clause( 'from' );
-		$this->clear_sql_clause( 'where_time' );
-		$this->clear_sql_clause( 'where' );
+		$this->clear_sql_clause( array( 'from', 'where_time', 'where' ) );
 
 		if ( isset( $query_args['before'] ) && '' !== $query_args['before'] ) {
 			if ( is_a( $query_args['before'], 'WC_DateTime' ) ) {
@@ -696,7 +628,7 @@ class DataStore {
 				$datetime_str = $query_args['before']->format( TimeInterval::$sql_datetime_format );
 			}
 			$sql_query['where_time_clause'] .= " AND {$table_name}.date_created <= '$datetime_str'";
-			$this->add_sql_clause( " AND {$table_name}.date_created <= '$datetime_str'", 'where_time' );
+			$this->add_sql_clause( 'where_time', " AND {$table_name}.date_created <= '$datetime_str'" );
 		}
 
 		if ( isset( $query_args['after'] ) && '' !== $query_args['after'] ) {
@@ -706,7 +638,7 @@ class DataStore {
 				$datetime_str = $query_args['after']->format( TimeInterval::$sql_datetime_format );
 			}
 			$sql_query['where_time_clause'] .= " AND {$table_name}.date_created >= '$datetime_str'";
-			$this->add_sql_clause( " AND {$table_name}.date_created >= '$datetime_str'", 'where_time' );
+			$this->add_sql_clause( 'where_time', " AND {$table_name}.date_created >= '$datetime_str'" );
 		}
 
 		return $sql_query;
@@ -733,7 +665,7 @@ class DataStore {
 		$sql_query['limit'] = "LIMIT {$sql_query['offset']}, {$sql_query['per_page']}";
 		// @todo remove $sql_query assignment after down stream classes are updated.
 		$this->clear_sql_clause( 'limit' );
-		$this->add_sql_clause( "LIMIT {$sql_query['offset']}, {$sql_query['per_page']}", 'limit' );
+		$this->add_sql_clause( 'limit', "LIMIT {$sql_query['offset']}, {$sql_query['per_page']}" );
 		return $sql_query;
 	}
 
@@ -816,7 +748,7 @@ class DataStore {
 		}
 		// @todo remove $sql_query return.
 		$this->clear_sql_clause( 'order_by' );
-		$this->add_sql_clause( $order_by_clause, 'order_by' );
+		$this->add_sql_clause( 'order_by', $order_by_clause );
 
 		return $sql_query;
 	}
@@ -835,9 +767,7 @@ class DataStore {
 			'where_clause'      => '',
 		);
 		// @todo remove $intervals_query after down stream classes are updated.
-		$this->clear_sql_clause( 'from' );
-		$this->clear_sql_clause( 'where_time' );
-		$this->clear_sql_clause( 'where' );
+		$this->clear_sql_clause( array( 'from', 'where_time', 'where' ) );
 
 		// @todo the array_merges can be removed leaving just the function calls once downstream is updated.
 		$intervals_query = array_merge( $intervals_query, $this->get_time_period_sql_params( $query_args, $table_name ) );
@@ -846,7 +776,7 @@ class DataStore {
 			$interval                         = $query_args['interval'];
 			$intervals_query['select_clause'] = TimeInterval::db_datetime_format( $interval, $table_name );
 			$this->clear_sql_clause( 'select' );
-			$this->add_sql_clause( TimeInterval::db_datetime_format( $interval, $table_name ), 'select' );
+			$this->add_sql_clause( 'select', TimeInterval::db_datetime_format( $interval, $table_name ) );
 		}
 
 		$intervals_query = array_merge( $intervals_query, $this->get_limit_sql_params( $query_args ) );
